Revert "[HUDI-6212] Hudi Spark 3.0.x integration  (#8714)"
This reverts commit 11aee9695ec30d8580758c8c81414ffd3c38f0b5.
