+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/spark/sql/hudi/TestMergeIntoTable.scala
@@ -114,65 +114,58 @@ class TestMergeIntoTable extends HoodieSparkSqlTestBase with ScalaAssertionSuppo
     })
   }
 
-  /**
-   * In Spark 3.0.x, UPDATE and DELETE can appear at most once in MATCHED clauses in a MERGE INTO statement.
-   * Refer to: `org.apache.spark.sql.catalyst.parser.AstBuilder#visitMergeIntoTable`
-   *
-   */
-  test("Test MergeInto with more than once update actions for spark >= 3.1.x") {
+  test("Test MergeInto with more than once update actions") {
+    withRecordType()(withTempDir {tmp =>
+      val targetTable = generateTableName
+      spark.sql(
+        s"""
+           |create table ${targetTable} (
+           |  id int,
+           |  name string,
+           |  data int,
+           |  country string,
+           |  ts bigint
+           |) using hudi
+           |tblproperties (
+           |  type = 'cow',
+           |  primaryKey = 'id',
+           |  preCombineField = 'ts'
+           | )
+           |partitioned by (country)
+           |location '${tmp.getCanonicalPath}/$targetTable'
+           |""".stripMargin)
+      spark.sql(
+        s"""
+           |merge into ${targetTable} as target
+           |using (
+           |select 1 as id, 'lb' as name, 6 as data, 'shu' as country, 1646643193 as ts
+           |) source
+           |on source.id = target.id
+           |when matched then
+           |update set *
+           |when not matched then
+           |insert *
+           |""".stripMargin)
+      spark.sql(
+        s"""
+           |merge into ${targetTable} as target
+           |using (
+           |select 1 as id, 'lb' as name, 5 as data, 'shu' as country, 1646643196 as ts
+           |) source
+           |on source.id = target.id
+           |when matched and source.data > target.data then
+           |update set target.data = source.data, target.ts = source.ts
+           |when matched and source.data = 5 then
+           |update set target.data = source.data, target.ts = source.ts
+           |when not matched then
+           |insert *
+           |""".stripMargin)
 
-    if (HoodieSparkUtils.gteqSpark3_1) {
-      withRecordType()(withTempDir { tmp =>
-        val targetTable = generateTableName
-        spark.sql(
-          s"""
-             |create table ${targetTable} (
-             |  id int,
-             |  name string,
-             |  data int,
-             |  country string,
-             |  ts bigint
-             |) using hudi
-             |tblproperties (
-             |  type = 'cow',
-             |  primaryKey = 'id',
-             |  preCombineField = 'ts'
-             | )
-             |partitioned by (country)
-             |location '${tmp.getCanonicalPath}/$targetTable'
-             |""".stripMargin)
-        spark.sql(
-          s"""
-             |merge into ${targetTable} as target
-             |using (
-             |select 1 as id, 'lb' as name, 6 as data, 'shu' as country, 1646643193 as ts
-             |) source
-             |on source.id = target.id
-             |when matched then
-             |update set *
-             |when not matched then
-             |insert *
-             |""".stripMargin)
-        spark.sql(
-          s"""
-             |merge into ${targetTable} as target
-             |using (
-             |select 1 as id, 'lb' as name, 5 as data, 'shu' as country, 1646643196 as ts
-             |) source
-             |on source.id = target.id
-             |when matched and source.data > target.data then
-             |update set target.data = source.data, target.ts = source.ts
-             |when matched and source.data = 5 then
-             |update set target.data = source.data, target.ts = source.ts
-             |when not matched then
-             |insert *
-             |""".stripMargin)
+      checkAnswer(s"select id, name, data, country, ts from $targetTable")(
+        Seq(1, "lb", 5, "shu", 1646643196L)
+      )
 
-        checkAnswer(s"select id, name, data, country, ts from $targetTable")(
-          Seq(1, "lb", 5, "shu", 1646643196L)
-        )
-      })
-    }
+    })
   }
 
   test("Test MergeInto with ignored record") {
